Module organization for "various_calculations".
